+++ BBS/Areas/Admin/Controllers/UserController.cs
@@ -7,6 +7,9 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using BBS.Data;
 using BBS.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using BBS.Controllers;
 
 namespace BBS.Areas.Admin.Controllers
 {
@@ -14,21 +17,31 @@ namespace BBS.Areas.Admin.Controllers
     public class UserController : Controller
     {
         private readonly BBSContext _context;
+        public UserManager<User> UserManager { get; }
 
-        public UserController(BBSContext context)
+        public UserController(BBSContext context, UserManager<User> userManager)
         {
             _context = context;
+            UserManager = userManager;
         }
 
         // GET: Admin/User
         public async Task<IActionResult> Index()
         {
+            if (!UserManager.GetUserName(User).ToLower().Equals("admin"))
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
             return View(await _context.Users.Where(a => a.UserName != "admin").ToListAsync());
         }
 
         // GET: Admin/User/Details/5
         public async Task<IActionResult> Details(string id)
         {
+            if (!UserManager.GetUserName(User).ToLower().Equals("admin"))
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -44,84 +57,13 @@ namespace BBS.Areas.Admin.Controllers
             return View(user);
         }
 
-        // GET: Admin/User/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: Admin/User/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Image,Introduce,AddTime,LastTime,State,Id,UserName,NormalizedUserName,Email,NormalizedEmail,EmailConfirmed,PasswordHash,SecurityStamp,ConcurrencyStamp,PhoneNumber,PhoneNumberConfirmed,TwoFactorEnabled,LockoutEnd,LockoutEnabled,AccessFailedCount")] User user)
-        {
-            if (ModelState.IsValid)
-            {
-                user.State = State.Offline;
-                user.AddTime = DateTime.Now;
-                _context.Add(user);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(user);
-        }
-
-        // GET: Admin/User/Edit/5
-        public async Task<IActionResult> Edit(string id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var user = await _context.Users.SingleOrDefaultAsync(m => m.Id == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
-            return View(user);
-        }
-
-        // POST: Admin/User/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(string id, [Bind("Image,Introduce,AddTime,LastTime,State,Id,UserName,NormalizedUserName,Email,NormalizedEmail,EmailConfirmed,PasswordHash,SecurityStamp,ConcurrencyStamp,PhoneNumber,PhoneNumberConfirmed,TwoFactorEnabled,LockoutEnd,LockoutEnabled,AccessFailedCount")] User user)
-        {
-            if (id != user.Id)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(user);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!UserExists(user.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(user);
-        }
-
         // GET: Admin/User/Delete/5
         public async Task<IActionResult> Delete(string id)
         {
+            if (!UserManager.GetUserName(User).ToLower().Equals("admin"))
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -142,6 +84,10 @@ namespace BBS.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
+            if (!UserManager.GetUserName(User).ToLower().Equals("admin"))
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
             var user = await _context.Users.SingleOrDefaultAsync(m => m.Id == id);
             _context.Users.Remove(user);
             await _context.SaveChangesAsync();
